Fix get marketMakers by collateral tokens
The selected market markets from the two options are wrong if the retrieve condition has the `fixedProductMarketMakers` array on different positions.
To fix this the market should fix with the collateral token associated, one market for the base token addres and other for the quote currency address.
"Yes" will be assume as the first option on the `outcomeIndex`.

- Add `setMarketMakers` method with `fixedProductMarketMakers` as parameter to set the right market makers on `baseTokenInfo` and `quoteCurrencyTokenInfo`.
- Add `getTokenPrice` method to calculate the token price by a given outcomeIndex.
- Add base token logo and `getLogoUrl` to get the token url based on the Safe app.
- Get `pairTokens` price from Unisap-v2 query in case the base token and quoteCurrency token pairs exists.
- Add `checksumAddress` and `fixedProductMarketMakers` to `baseTokenInfo` and `quoteCurrencyTokenInfo` json object.